# microservicios/src/main/resources/META-INF/openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Pago",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:8080/",
			"description": "Servidor de desarrollo."
		}
	],
	"paths": {
		"/pago": {
			"post": {
				"requestBody": {
					"description": "Se envia el pago",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pago"
							}
						}
					}
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pago"
								}
							}
						},
						"description": "La respuesta 201 indica que se ha creado el pago con éxito."
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"codigoError": {
											"type": "string"
										},
										"mensajeError": {
											"type": "string"
										}
									},
									"example": {
										"codigoError": "CRP_422",
										"mensajeError": "Error en el cuerpo de la solicitud"
									}
								}
							}
						},
						"description": "Se responde con código 422 en caso de que un parámetro esté mal definido"
					}
				}
			}
		},
		"/pago/{id}": {
			"get": {
				"description": "Retorna la información del pago correspondiente al ID indicado.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "ID del pago a consultar",
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pago"
								}
							}
						},
						"description": "La respuesta 200 indica que si existe el registro."
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"codigoError": {
											"type": "string"
										},
										"mensajeError": {
											"type": "string"
										},
										"resultado": {
											"type": "string"
										}
									}
								}
							}
						},
						"description": "Se responde con código 404 en caso de que no se encuentre el pago"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Pago": {
				"type": "object",
				"properties": {
					"clienteId": {
						"type": "string"
					},
					"metodoPagoId": {
						"type": "string"
					},
					"monto": {
						"type": "number",
						"format": "double"
					},
					"refVenta": {
						"type": "string"
					},
					"estadoId": {
						"type": "integer",
						"format": "int32"
					},
					"voucherId": {
						"type": "string"
					}
				},
				"requeried": [
					""
				],
				"example": {
					"clienteId": 5602,
					"metodoPagoId": 1,
					"monto": 50000.00,
					"refVenta": "JKasaihsLDSO7"
				}
			}
		}
	}
}